+++ app/common/util/src/main/java/io/syndesis/common/util/openapi/CapturingCodec.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.server.api.generator.swagger.util;
+package io.syndesis.common.util.openapi;
 
 import java.io.IOException;
 import java.util.Iterator;
