+++ app/server/api-generator/pom.xml
@@ -43,11 +43,6 @@
 
     <dependency>
       <groupId>io.swagger</groupId>
-      <artifactId>swagger-core</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>io.swagger</groupId>
       <artifactId>swagger-parser</artifactId>
     </dependency>
